Rework ConfigurationContext to use an Enum to read configuration properties
